Optimizers

+++ include/FenDL/Optimizers/ADAM.hxx
@@ -4,7 +4,7 @@
 #include <FenDL/Optimizer.hxx>
 
 class ADAM : public Optimizer{
-    void updateWeights(Matrixd& weights,Matrixd& gradient,double ls,int epoch) override;
+    void updateWeights(Matrixd& weights,Matrixd& gradient,double ls,double epoch) override;
 };
